CSR changed to SSG
