+++ src/components/Person/Person.tsx
@@ -1,45 +1,18 @@
-import { FC, useEffect, useState } from 'react'
+import { FC } from 'react'
 
-import defaultData from './data_mock.json'
 import styles from './Person.module.scss'
 import { PersonFilmography } from './PersonFilmography'
 import { PersonHeader } from './PersonHeader'
 
-export interface IFilmId {
-  film_id: string
-}
-
-export interface IPerson {
-  person_id: string
-  first_name_ru: string
-  last_name_ru: string
-  first_name_en: string
-  last_name_en: string
-  img: string
-  films: IFilmId[]
-}
-
 export interface PersonProps {
-  person_id: string
-  pathDataSrc: string
+  person: any
   maxShowFilms: number
 }
 
 export const Person: FC<PersonProps> = ({
-  person_id,
-  pathDataSrc,
+  person,
   maxShowFilms,
 }) => {
-  const [person, setPerson] = useState(defaultData as IPerson)
-
-  useEffect(() => {
-    const fetchData = async () => {
-      const response = await fetch(`${pathDataSrc}${person_id}`)
-      const data = await response.json()
-      setPerson(data)
-    }
-    fetchData()
-  }, [person_id, pathDataSrc])
 
   return (
     <div className={styles.person}>
